pyproject.toml
[project]
name = "footstep-m2"
version = "0.1.0"
description = "足音自己位置推定の頑健化パイプライン（修士論文）"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "torchaudio>=2.2",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "pillow>=10.0",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "tensorboard>=2.16",
    "librosa>=0.10",
    "soundfile>=0.12",
]
